Dropping no-restricted-imports rule
In https://github.com/paritytech/opstooling/pull/100, it was suggested
to disallow imports from upper levels. The rule was never respected as a
design choice, and mostly was either bypassed by absolute imports, or
disabled.
My opinion is that it would be better to drop it altogether, and allow
all imports.